asta_data_crawler/requirements.txt
functions-framework==3.*
supabase==2.*
pandas==2.*
numpy==1.*
requests==2.*
beautifulsoup4==4.*
xgboost==2.*
scikit-learn==1.*
python-dotenv==1.*
google-cloud-secret-manager==2.*
vertexai>=1.0.0
google-cloud-aiplatform>=1.38.1
google-api-python-client>=2.0.0
google-cloud-translate>=3.0.0,<4.0.0
google-auth>=2.0.0,<3.0.0
google-auth-oauthlib>=1.0.0,<2.0.0
google-auth-httplib2>=0.1.0,<1.0.0
youtube-transcript-api>=1.2.2,<2.0.0
# Additions for Crawl4AI + Groq scraping template
crawl4ai>=0.7.4,<1.0.0
groq>=0.9.0,<1.0.0
google-cloud-translate>=3.0.0,<4.0.0 # Already added, good for potential future use
